# improve performance of join by make it internal