Avoids exponential string append

Before:
BenchmarkJoinShort
BenchmarkJoinShort-4       26947             44399 ns/op
BenchmarkJoinLong
BenchmarkJoinLong-4          145           7295904 ns/op

After:
BenchmarkJoinShort
BenchmarkJoinShort-4   	   81789	     14456 ns/op
BenchmarkJoinLong
BenchmarkJoinLong-4    	   22002	     53791 ns/op