@wader 2021-09-20
Good question, yes should work, internal functions does not know about generators, will only get values. But i'm not sure how it get compiled to make it work in these cases, now i got curious. So i guess a call to a function that has any argument that results in no outputs will itself result in no output? seems resonable.

Tried your examples with the change:
```sh
$ go run cmd/gojq/main.go -nr '["a","b"] | join(",", "\n")'
a,b
a
b
$ go run cmd/gojq/main.go -nr '["a","b"] | join(empty)'
```