@wader 2021-09-20
Maybe it's possible to have a more generic internal help function for improving performance of reduce + binop functions?  like keep map over array in jq and then pass result to a internal function with argument for how it should aggregate?