@itchyny 2021-09-21
You don't need to worry about stream as argument of internal functions because they are processed before called.
Don't break `{x:1,y:2} | join(",")`.