@itchyny 2021-09-21
This growing complexity is the reason I stopped making it internal. I don't like the duplicates with the executor and pathValue to be used by internal functions. So how about `def join($x): [.[]] | _join($x);` and making `_join` internal? Or `if type != "array" then [.[]] end | _join($x)` if we avoid cloning arrays.