@wader 2021-09-21
Updated to use own string builder, long when from 57242 to 48389