@itchyny 2021-09-21
> By avoid cloning the array do you mean implement our own version of strings.Join so that we don't have to copy it?

No, use `strings.Join`. I just mean `[.[]] | _join($x)` clones the input array.