@wader 2021-09-21
Ok switching back to `strings.Join`